// src/app/shared/models/kPISummaryDashboard.model.ts
export interface IKPISummaryDashboard {
    RInhScruJobTtlTime: number;
    RInhPrcoJobTtlTime: number;
    RInhValJobTtlTime: number;
    RCompScruJobTtlTime: number;
    RCompPrcoJobTtlTime: number;
    RCompValJobTtlTime: number;
    RCompletedScruJobCount: number;
    RInHandScruJobCount: number;
    RCompletedProcJobCount: number;
    RInHandProcJobCount: number;
    RCompletedValiJobCount: number;
    RInHandValiJobCount: number;
    RAvgInhScruJobTtlTime: number;
    RAvgInhPrcoJobTtlTime: number;
    RAvgInhValJobTtlTime: number;
    RAvgCompScruJobTtlTime: number;
    RAvgCompPrcoJobTtlTime: number;
    RAvgCompValJobTtlTime: number;
    RSumryCompJobCount: number;
    RSumryCompTtlTime: number;
    RSumryInhJobCount: number;
    RSumryInhTtlTime: number;
    RSumryAvgCompJobCount: number;
    RSumryAvgInhTtlTime: number;
    RSumryCompJobCountWw: number;
    RSumryCompTtlTimeWw: number;
    RSumryInhJobCountWw: number;
    RSumryInhTtlTimeWw: number;
    RSumryAvgInhJobCountWw: number;
    RSumryAvgInhTtlTimeWw: number;
}